backend/requirements.txt
google-generativeai
yfinance
matplotlib
pandas
python-dotenv
flask
flask-cors